+++ Client/src/model/Coordinate.java
@@ -4,14 +4,11 @@ import java.io.Serializable;
 
 public class Coordinate implements Serializable{
 	
-	/**
-	 * 
-	 */
 	private static final long serialVersionUID = -4239112052905532281L;
 	
-	public int x, y;
+	public double x, y;
 	
-	public Coordinate(int x, int y) {
+	public Coordinate(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
