+++ Server/src/Coordinate.java
@@ -2,14 +2,11 @@ import java.io.Serializable;
 
 public class Coordinate implements Serializable {
 	
-	/**
-	 * 
-	 */
 	private static final long serialVersionUID = 3869446341373817110L;
 	
-	public int x, y;
+	public double x, y;
 	
-	public Coordinate(int x, int y) {
+	public Coordinate(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
